lexer update
previously when the lexer found a match, it adjusted the position it looked for matches in to the end of the last match.  Updated it so that is instead deletes matched text from the code and searches again from the beginning of the updated code. This allows a lstrip() call to remove whitespace which caused intermittent issues on code with unexpected whitespace